Converts the local-only version of the app to use AngularJS instead of jQuery. Adds config file lookups for the API hooks for the server-based version.

+++ clearTimer.php
@@ -1,5 +1,12 @@
 <?php
-$timeFile = "time_ending.txt";
-unlink($timeFile);
+$configs = parse_ini_file('default.ini.php', true);
+if (!is_array($configs)) {
+    die(header("Status: 501"));
+}
+
+$timeFile = $configs['paths']['time_file'];
+if (file_exists($timeFile)) {
+    unlink($timeFile);
+}
 
 exit(header("Status: 200"));
